Struct cancellation::CancellationTokenSource

pub struct CancellationTokenSource {
    // some fields omitted
}

Methods

impl CancellationTokenSource

fn new() -> CancellationTokenSource

fn token(&self) -> &Arc<CancellationToken>

Gets the token managed by this CancellationTokenSource.

The Arc can be cloned so that the token can be passed into asynchronous tasks that might outlive the CancellationTokenSource.

fn cancel(&self)

Marks the cancellation token as canceled.

Executes the on_cancel callback of any active CancellationToken::run invocations. Has no effect if the cancellation token was already canceled.

fn cancel_after(&self, dur: Duration)

Creates a new, detached thread that waits for the specified duration and then marks the cancellation token as canceled.